requirements.txt
# --- Core API & Web ---
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic>=2.9.0
pydantic-settings>=2.2.1

# --- Database & Persistence ---
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# --- Data Science & Analysis ---
numpy==1.26.4
pandas==2.2.1
pyarrow==15.0.0
fastparquet==2024.2.0
tqdm==4.66.2
ta==0.11.0
PyWavelets
statsmodels

# --- Machine Learning (Elite 10 Ensemble) ---
scikit-learn==1.4.1.post1
xgboost==2.0.3
catboost==1.2.3
lightgbm==4.3.0
joblib==1.3.2

# --- Utilities & Logging ---
python-dotenv==1.0.1
requests==2.31.0
loguru==0.7.2
httpx==0.27.0

# --- Event Streaming (Kafka) ---
aiokafka==0.12.0
kafka-python-ng==2.2.3

# --- Institutional Additions ---
ujson==5.9.0
PyYAML==6.0.1
prometheus-client==0.20.0
smartapi-python==1.3.4
websocket-client
pyotp
cachetools
tenacity
yfinance==0.2.37
feedparser==6.0.11
textblob==0.17.1
nsepython
